Get rid of `MemoryLoginsStorage.kt`
Fixes #2310
